refactor: Lotto class에 랜던 번호 생성 추가
