+++ src/view/output.js
@@ -1,10 +1,10 @@
 import { PRIZE } from "../constants/constant.js";
 
 export const OutputView = {
-  outputLottoNumber(lottos) {
-    console.log(lottos.length, "개를 구매했습니다.");
-    for (let i = 0; i < lottos.length; i++) {
-      console.log(lottos[i].getNumber());
+  outputLottoNumber(randomLottos) {
+    console.log(randomLottos.length, "개를 구매했습니다.");
+    for (let i = 0; i < randomLottos.length; i++) {
+      console.log(randomLottos[i].getNumber());
     }
   },
